Fix all enabled GCC, Clang, and cppcheck warnings
This corrects the warnings generated by the compilers (GCC, Clang)
as well as the static analysis tool cppcheck. It also re-enables some
warnings that were previously surpressed in CMakeLists although most
of these did not generate any new messages. This commit also updates
the rebuild.sh script such that static analysis is only run if the
code successfully compiles.

+++ src/HS.cpp
@@ -117,8 +117,8 @@ RecordPtr HS::promptForRecord()
 bool HS::sendRecord(const RecordPtr& r)
 {
   auto addr = Config::getAuthority()[0];
-  auto socks = SocksClient::getCircuitTo(addr["ip"].asString(),
-                                         addr["port"].asInt(), 9050);
+  auto socks = SocksClient::getCircuitTo(
+      addr["ip"].asString(), static_cast<short>(addr["port"].asInt()), 9050);
   if (!socks)
     throw std::runtime_error("Unable to connect!");
